+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master (unreleased)
 
+- Add ability to incrementally configure a paginator
+
 - Add ability to get the total number of records
 
     ```ruby